refactor(cli-max.ts): extract out the interfaces related to command
